feat: Implement lazy loading for NFT images
This commit adds lazy loading functionality for NFT images in the `metamask.js` file. The previous code was commented out, and a new API call using the `query` function from `api.js` was added to fetch the NFT image link. The fetched link is then used to mint the NFT for the user's account.
